Added checksum to bvh analysis.
It creates a checksum of the vertex ids of the first triangle in each leafnodes. If the checksum of two bvhs is the same then they all have the same leafnodes with the same primitives.
Most scenes have the same checksum for the same leafsize (when sorting each split), but its not guaranteed